Fixing the warning messages captured under mac os # Consider using `git commit -m 'One line title' && arc diff`. # You will save time by running lint and unit in the background.
Summary: The work to make sure mac os compiles rocksdb is not completed yet. But at least we can start cleaning some warnings captured only by g++ from mac os..

Test Plan: ran make in mac os

Reviewers: dhruba

Reviewed By: dhruba

CC: leveldb

Differential Revision: https://reviews.facebook.net/D14049

+++ db/db_test.cc
@@ -3644,7 +3644,7 @@ TEST(DBTest, SnapshotFiles) {
       char buffer[4096];
       Slice slice;
       while (size > 0) {
-        uint64_t one = std::min(sizeof(buffer), size);
+        uint64_t one = std::min(uint64_t(sizeof(buffer)), size);
         ASSERT_OK(srcfile->Read(one, &slice, buffer));
         ASSERT_OK(destfile->Append(slice));
         size -= slice.size();
